% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_container}
\alias{docker_container}
\title{Management commands for working with a particular docker container}
\description{
Methods for working with a particular docker container.  Container
objects are returned by creating or running a docker container, or
by using \code{$containers$get} to fetch an existing container by
name or id.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("docker_container")}
}
\seealso{
\code{\link{docker_container_collection}} for other
  container management methods.
}
