% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_container_collection}
\alias{docker_container_collection}
\title{Management commands for working with docker containers}
\description{
Methods for working with docker containers.  This object is
  \code{$container} within a \code{\link{docker_client}} object.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("container")}
}
\seealso{
\code{\link{docker_container}} for information on
  container objects.
}
