% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_exec}
\alias{docker_exec}
\title{Commands for working with a docker exec instance}
\description{
Methods for working with docker "exec" instances, which are
returned by running \code{exec} on a running container
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("docker_exec")}
}
\seealso{
\code{\link{docker_container}}
}
