% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_image}
\alias{docker_image}
\title{Management commands for working with a particular docker image}
\description{
Methods for working with a particular docker image.  Image objects
are returned by building or pulling a docker image, or by using
\code{$images$get} to fetch an existing image by name or id.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("docker_image")}
}
\seealso{
\code{\link{docker_image_collection}} for other
  image management methods.
}
