% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_image_collection}
\alias{docker_image_collection}
\title{Management commands for working with docker images}
\description{
Methods for working with docker images.  This object is
  \code{$images} within a \code{\link{docker_client}} object.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("images")}
}
\seealso{
\code{\link{docker_image}} for information on
  image objects.
}
