% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_network}
\alias{docker_network}
\title{Management commands for working with a particular docker network}
\description{
Methods for working with a particular docker network.  Network
objects are returned by creating a docker network, or by using
\code{$network$get} to fetch an existing network by name or id.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("docker_network")}
}
\seealso{
\code{\link{docker_network_collection}} for other
  network management methods.
}
