% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_network_collection}
\alias{docker_network_collection}
\title{Management commands for working with docker networks}
\description{
Methods for working with docker networks.  This object is
  \code{$networks} within a \code{\link{docker_client}} object.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("networks")}
}
\seealso{
\code{\link{docker_network}} for information on
  network objects.
}
