% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_node}
\alias{docker_node}
\title{Management commands for working with a particular docker node}
\description{
Methods for working with a particular docker node.  Node objects
are by using \code{$node$get} to fetch an existing node by name or
id.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("docker_node")}
}
\seealso{
\code{\link{docker_node_collection}} for other
  node management methods.
}
