% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_node_collection}
\alias{docker_node_collection}
\title{Management commands for working with swarm nodes}
\description{
Methods for managing docker swarm nodes.  This object is
  \code{$node} within a \code{\link{docker_client}} object.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("node")}
}
\seealso{
\code{\link{docker_swarm_collection}} for management
  commands for the swarm itself.
}
