% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_plugin}
\alias{docker_plugin}
\title{Management commands for working with a particular docker plugin}
\description{
Methods for working with a particular docker plugin.  Plugin
objects are returned by installing or building a docker plugin, or
by using \code{$plugins$get} to fetch an existing plugin by name
or id.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("docker_plugin")}
}
\seealso{
\code{\link{docker_plugin_collection}} for other
  plugin management methods.
}
