% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_plugin_collection}
\alias{docker_plugin_collection}
\title{Management commands for working with docker plugins}
\description{
Methods for working with docker plugins.  This object is
  \code{$plugin} within a \code{\link{docker_client}} object.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("plugin")}
}
\seealso{
\code{\link{docker_plugin}} for information on
  plugin objects.
}
