% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_secret_collection}
\alias{docker_secret_collection}
\title{Management commands for working with swarm secrets}
\description{
Methods for managing docker swarm secrets.  This object is
  \code{$secrets} within a \code{\link{docker_client}} object.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("secrets")}
}
\seealso{
\code{\link{docker_swarm}} for management commands for
  the swarm itself, and \code{\link{docker_config_collection}} for
  a similar inteface for configuring non-sensitive configurations.
}
