% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_service}
\alias{docker_service}
\title{Management commands for working with a particular docker service}
\description{
Methods for working with a particular docker service.  Service
objects are returned by creating a docker service, or by using
\code{$service$get} to fetch an existing service by name or id.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("docker_service")}
}
\seealso{
\code{\link{docker_service_collection}} for other
  service management methods.
}
