% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_swarm_collection}
\alias{docker_swarm_collection}
\title{Management commands for working with docker swarm}
\description{
Methods for managing the docker swarm.  This object is
  \code{$swarm} within a \code{\link{docker_client}} object.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("swarm")}
}
\seealso{
\code{\link{docker_node}} for management commands for
  swarm nodes.
}
