% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_task_collection}
\alias{docker_task_collection}
\title{Management commands for working with docker tasks}
\description{
Methods for working with docker tasks.  This object is
  \code{$task} within a \code{\link{docker_client}} object.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("task")}
}
\seealso{
\code{\link{docker_task}} for information on
  task objects.
}
