% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_volume_collection}
\alias{docker_volume_collection}
\title{Management commands for working with docker volumes}
\description{
Methods for working with docker volumes.  This object is
  \code{$volume} within a \code{\link{docker_client}} object.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("volume")}
}
\seealso{
\code{\link{docker_volume}} for information on
  volume objects.
}
