% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stevedore.R
\docType{package}
\name{stevedore}
\alias{stevedore}
\alias{stevedore-package}
\title{Docker Client For R}
\description{
A docker client for R, using \code{curl} and an automatically
generated interface.  To get started, please see the package
vignette (\code{vignette("stevedore")} will work if the package
was installed with the vignettes).
}
