% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_available.R
\name{docker_available}
\alias{docker_available}
\title{Test if docker available}
\usage{
docker_available(..., verbose = FALSE)
}
\arguments{
\item{...}{Passed through to \code{docker_client} (e.g.,
\code{api_version}, \code{host}).}

\item{verbose}{Logical, indicating if information should be
printed about failures to connect.  If \code{FALSE} (the
default) the function runs silently.}
}
\value{
Logical scalar, \code{TRUE} if
  \code{\link{docker_client}(...)} would succeed.
}
\description{
Test if we can construct a docker client and confirm that we can
communicate with it.  This is intended to help in debug connection
issues, and also for use in tests.  For example, you might
implement a testthat skip test that skips if
\code{stevedore::docker_available()} returns \code{FALSE} to
conditionally use stevedore/docker within tests.
}
\details{
Reasons for failure to connect might include:

\itemize{
\item You do not have a docker daemon running

\item You have docker installed but the socket in a nonstandard
place and have not adjusted environment variables accordingly

\item You do not have permission to write to the docker socket

\item You are on windows and the required python packages to get
  everything working there are not present or configured correctly

\item There are problems arranging verification over https/tls.
}

If \code{versose} is \code{TRUE} then some diagnostic information
will be printed.
}
\examples{
# Is docker available on your system?
stevedore::docker_available()
}
