% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_available.R
\name{docker_available}
\alias{docker_available}
\title{Test if docker available}
\usage{
docker_available(api_version = NULL, url = NULL, ...,
  http_client_type = NULL, verbose = FALSE)
}
\arguments{
\item{api_version}{Version of the API request from the api.
Options are \code{NULL} (the default) - use the package's
default version (currently
\Sexpr{stevedore:::DOCKER_API_VERSION_DEFAULT}), a version as a
string or \code{\link{numeric_version}} object (supported
between \Sexpr{stevedore:::DOCKER_API_VERSION_MIN} and
\Sexpr{stevedore:::DOCKER_API_VERSION_MAX}), or the string
\code{detect} which will use the highest version out of the
version reported by the api and
\Sexpr{stevedore:::DOCKER_API_VERSION_MAX}.}

\item{url}{The URL for the docker daemon.  This can be an absolute
file path (for a unix socket on macOS/Linux), a named pipe
(e.g., \code{npipe:////./pipe/docker_engine}) on Windows, or
(eventually) an http or https url (e.g.,
\code{https://localhost:8888}), though this is not yet
supported.}

\item{...}{Reserved for future use.  Passing in any unrecognised
argument will throw an error.}

\item{http_client_type}{HTTP client type to use.  The options are
(currently) "curl", which uses the \code{curl} package (works
over unix sockets and eventually over TCP) and \code{httppipe}
which works over unix sockets and eventually windows named
pipes, using the Docker SDK's pipe code via the \code{httppipe}
package.  Not all functionality is supported with the
\code{httppipe} client.  This option may eventually be moved
into the \code{...} argument as is not intended for end-user
use; it is primarily intended for debugging in development
(forcing the \code{httppipe} client where the \code{curl} client
would ordinarily be preferred).}

\item{verbose}{Logical, indicating if information should be
printed about failures to connect.  If \code{FALSE} (the
default) the function runs silently.}
}
\value{
Logical scalar, \code{TRUE} if
  \code{\link{docker_client}(...)} would succeed.
}
\description{
Test if we can construct a docker client and ping it.  This is
intended to help in debug connection issues, and also for use in
tests.  For example, you might implement a testthat skip test that
skips if \code{stevedore::docker_available()} returns \code{FALSE}
to conditionally use stevedore/docker within tests.
}
\details{
Reasons for failure to connect might include:

\itemize{
\item You do not have docker installed

\item You have docker installed but the socket in a nonstandard
place and have not adjusted environment variables accordingly

\item You do not have permission to write to the docker socket

\item You are on windows and the required python packages to get
  everything working there are not present or configured correctly.
}
}
\examples{
# Is docker available on your system?
stevedore::docker_available()
}
