% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_client.R
\name{docker_client}
\alias{docker_client}
\title{Create docker client}
\usage{
docker_client(..., api_version = NULL)
}
\arguments{
\item{...}{Reserved for future use}

\item{api_version}{Version of the API request from the api.
Options are \code{NULL} (the default) - use the package's
default version (currently 1.29), a version as a string or
\code{\link{numeric_version}} object (supported between 1.25 and
1.32), or the string \code{detect} which will use the highest
version out of the version reported by the api and 1.32}
}
\description{
Create a docker client object
}
