% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_client.R
\name{docker_client}
\alias{docker_client}
\title{Create docker client}
\usage{
docker_client(..., api_version = NULL, host = NULL, cert_path = NULL,
  tls_verify = NULL, http_client_type = NULL, quiet = FALSE,
  ignore_environment = FALSE)
}
\arguments{
\item{...}{Reserved for future use.  Passing in any unrecognised
argument will throw an error.  Part of the role of this argument
is to force use of named arguments until the API is stabilised.}

\item{api_version}{Version of the API request from the api.
Options are \code{NULL} (the default) - use the package's
default version (currently
\Sexpr{stevedore:::DOCKER_API_VERSION_DEFAULT}), a version as a
string or \code{\link{numeric_version}} object (supported
between \Sexpr{stevedore:::DOCKER_API_VERSION_MIN} and
\Sexpr{stevedore:::DOCKER_API_VERSION_MAX}), or the string
\code{detect} which will use the highest version out of the
version reported by the api and
\Sexpr{stevedore:::DOCKER_API_VERSION_MAX}.  If the environment
variable \code{DOCKER_API_VERSION} is set then this will be used
as the version (this is the same behaviour as the command line
client).}

\item{host}{The URL for the docker daemon.  This can be a unix
socket (e.g., unix:///var/run/docker.sock) on macOS/Linux, a
named pipe (e.g., \code{npipe:////./pipe/docker_engine}) on
Windows, or an http or https url (e.g.,
\code{https://localhost:2376}).  If not given, we use the
environment variable \code{DOCKER_HOST}, falling back on the
default socket or named pipe (for macOS/unix and windows
respectively).}

\item{cert_path}{The path to a directory containing certificate
files.  If using an \code{https} url this is required.  If not
given, we use the environment variable \code{DOCKER_CERT_PATH}.
This is ignored without warning if used with a socket or named
pipe connection.}

\item{tls_verify}{Logical, indicating if TLS should be verified.
This is only used if using an https connction (i.e., host is a
tcp/http/https url and\code{cert_path} is given).  If not given,
we use the environment variable \code{DOCKER_TLS_VERIFY}.}

\item{http_client_type}{HTTP client type to use.  The options are
(currently) "curl", which uses the \code{curl} package (works
over unix sockets and over TCP) and \code{httppipe} which works
over unix sockets and windows named pipes, using the Docker
SDK's pipe code via the \code{httppipe} package.  Not all
functionality is supported with the \code{httppipe} client.
This option may eventually be moved into the \code{...} argument
as is not intended for end-user use; it is primarily intended
for debugging in development (forcing the \code{httppipe} client
where the \code{curl} client would ordinarily be preferred).}

\item{quiet}{Suppress informational messages.}

\item{ignore_environment}{Logical, indicating if environment
variables (\code{DOCKER_HOST}, \code{DOCKER_CERT_PATH},
\code{DOCKER_TLS_VERIFY} and \code{DOCKER_API_VERSION}) should
be ignored.}
}
\description{
Create a docker client object
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help()}
}
