% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_client.R
\name{docker_client}
\alias{docker_client}
\title{Create docker client}
\usage{
docker_client(api_version = NULL, url = NULL, ...,
  http_client_type = NULL)
}
\arguments{
\item{api_version}{Version of the API request from the api.
Options are \code{NULL} (the default) - use the package's
default version (currently 1.29), a version as a string or
\code{\link{numeric_version}} object (supported between 1.25 and
1.33), or the string \code{detect} which will use the highest
version out of the version reported by the api and 1.33}

\item{url}{The URL for the docker daemon.  This can be an absolute
file path (for a unix socket on macOS/Linux), a named pipe
(e.g., \code{npipe:////./pipe/docker_engine}) on Windows, or
(eventually) an http or https url (e.g.,
\code{https://localhost:8888}), though this is not yet
supported.}

\item{...}{Reserved for future use.  Passing in any unrecognised
argument will throw an error.}

\item{http_client_type}{HTTP client type to use.  The options are
(currently) "curl", which uses the \code{curl} package (works
over unix sockets and eventually over TCP) and \code{httppipe}
which works over unix sockets and eventually windows named
pipes, using the Docker SDK's pipe code via the \code{httppipe}
package.  Not all functionality is supported with the
\code{httppipe} client.  This option may eventually be moved
into the \code{...} argument as is not intended for end-user
use; it is primarily intended for debugging in development
(forcing the \code{httppipe} client where the \code{curl} client
would ordinarily be preferred).}
}
\description{
Create a docker client object
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help()}
}
