% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_config_collection}
\alias{docker_config_collection}
\title{Management commands for working with swarm configs}
\description{
Methods for managing docker swarm configs.  This object is
  \code{$configs} within a \code{\link{docker_client}} object.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("configs")}
}
\seealso{
\code{\link{docker_swarm_collection}} for management
  commands for the swarm itself, and
  \code{\link{docker_secret_collection}} for a similar inteface
  for configuring sensitive configurations.
}
