% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{docker_service_collection}
\alias{docker_service_collection}
\title{Management commands for working with docker services}
\description{
Methods for working with docker services.  This object is
  \code{$services} within a \code{\link{docker_client}} object.
}
\details{
\Sexpr[results=rd,stage=render]{stevedore:::generate_help("services")}
}
\seealso{
\code{\link{docker_service}} for information on
  service objects.
}
